Added typescript configuration
