--- user-management-service/src/routes/test-route.js
@@ -1,9 +1,0 @@
-import express from "express";
-
-const router = express.Router();
-
-router.get("/", (req, res) => {
-  res.send("hello from tst route - product catalouge");
-});
-
-export { router as testRoute };
