+++ user-management-service/src/routes/test-route.ts
@@ -1,0 +1,9 @@
+import express, { Request, Response } from "express";
+
+const router = express.Router();
+
+router.get("/", (req: Request, res: Response) => {
+  res.send("hello from tst route - users");
+});
+
+export { router as testRoute };
